# api/openapi.yaml
openapi: 3.1.0
info:
  title: Skyline MCP Gateway API
  description: >-
    REST and MCP protocol API for the Skyline MCP Gateway server. Manages API
    profiles, tool discovery, tool execution, credential verification, admin
    observability, and OAuth 2.1 flows for ChatGPT MCP compatibility.
  version: 0.9.16
  license:
    name: AGPL-3.0
    url: https://www.gnu.org/licenses/agpl-3.0.html
servers:
  - url: https://localhost:8191
    description: Default local development server
...
security: []
paths:
...
  /detect:
    post:
      operationId: detectAPIs
      summary: Auto-detect API spec formats at a base URL
      description: >-
        Probes common spec paths (OpenAPI, Swagger, GraphQL, WSDL, OData, etc.)
        and returns which ones were found. Rate-limited to 5 requests per minute.
      tags: [detection]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DetectRequest'
      responses:
        '200':
          description: Detection results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DetectResponse'
        '400':
          description: Invalid request body or missing base_url
        '429':
          description: Rate limited
  /test:
    post:
      operationId: testSpecURL
      summary: Test reachability of a spec URL
      tags: [detection]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TestRequest'
      responses:
        '200':
          description: Reachability test result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TestResponse'
        '400':
          description: Invalid request body or missing spec_url
  /operations:
    post:
      operationId: listOperations
      summary: Fetch and parse a spec, return its operations
      tags: [detection]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OperationsRequest'
      responses:
        '200':
          description: Parsed operations list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationsResponse'
        '400':
          description: Invalid request body or missing spec_url

  # ──────────────────────────────────────────────
  # Verification
  # ──────────────────────────────────────────────
  /verify:
    post:
      operationId: verifyCredentials
      summary: Verify credentials against a third-party service
      description: >-
        Proxies a credential verification request to Slack, GitLab, Jira, or Gmail.
        Needed because browsers cannot call third-party APIs directly (CORS).
      tags: [verification]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyRequest'
      responses:
        '200':
          description: Verification result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VerifyResponse'
        '400':
          description: Invalid request or unsupported service

  # ──────────────────────────────────────────────
  # Admin
  # ──────────────────────────────────────────────
  /admin/auth:
    get:
      operationId: checkAdminSession
      summary: Check if the current admin session cookie is valid
      tags: [admin]
      security:
        - AdminSession: []
      responses:
        '200':
          description: Session is valid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusOk'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      operationId: adminLogin
      summary: Authenticate with admin token, receive session cookie
      tags: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token]
              properties:
                token:
                  type: string
                  description: Admin token from config.yaml or server startup
      responses:
        '200':
          description: Login successful; skyline_admin cookie is set
          headers:
            Set-Cookie:
              description: >-
                skyline_admin session cookie (Secure, HttpOnly, SameSite=Strict, 7-day max-age)
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusOk'
        '400':
          description: Invalid request body
        '401':
          description: Invalid admin token
...
  /admin/audit:
    get:
      operationId: getAuditLog
      summary: Query audit log entries
      tags: [admin]
      security:
        - AdminSession: []
      parameters:
        - name: profile
          in: query
          description: Filter by profile name
          schema:
            type: string
        - name: event_type
          in: query
          description: Filter by event type (execute, connect, disconnect, error)
          schema:
            type: string
        - name: tool_name
          in: query
          description: Filter by tool name
          schema:
            type: string
        - name: limit
          in: query
          description: Maximum number of events to return (default 100, max 1000)
          schema:
            type: integer
            default: 100
            maximum: 1000
      responses:
        '200':
          description: Audit log events
          content:
            application/json:
              schema:
                type: object
                required: [events, count]
                properties:
                  events:
                    type: array
                    items:
                      $ref: '#/components/schemas/AuditEvent'
                  count:
                    type: integer
        '401':
          $ref: '#/components/responses/Unauthorized'
  /admin/stats:
    get:
      operationId: getStats
      summary: Aggregated statistics (audit + metrics)
      tags: [admin]
      security:
        - AdminSession: []
      parameters:
        - name: profile
          in: query
          description: Filter stats by profile name
          schema:
            type: string
        - name: since
          in: query
          description: Start of time window (RFC 3339). Defaults to 24 hours ago.
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Combined audit stats and metrics snapshot
          content:
            application/json:
              schema:
                type: object
                properties:
                  audit_stats:
                    $ref: '#/components/schemas/AuditStats'
                  metrics_snapshot:
                    $ref: '#/components/schemas/MetricsSnapshot'
                  period:
                    type: object
                    properties:
                      since:
                        type: string
                        format: date-time
                      until:
                        type: string
                        format: date-time
        '401':
          $ref: '#/components/responses/Unauthorized'
  /admin/config:
    get:
      operationId: getConfig
      summary: Read server configuration (config.yaml)
      tags: [admin]
      security:
        - AdminSession: []
      responses:
        '200':
          description: Server configuration (raw YAML + parsed sections)
          content:
            application/json:
              schema:
                type: object
                properties:
                  raw:
                    type: string
                    description: Raw YAML content of config file
                  server:
                    type: object
                    additionalProperties: true
                  runtime:
                    type: object
                    additionalProperties: true
                  audit:
                    type: object
                    additionalProperties: true
                  profiles:
                    type: object
                    additionalProperties: true
                  security:
                    type: object
                    additionalProperties: true
                  logging:
                    type: object
                    additionalProperties: true
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      operationId: saveConfig
      summary: Save server configuration (raw YAML body)
      description: >-
        Accepts raw YAML in the request body. Validates YAML syntax before saving.
        Requires server restart for changes to take effect.
      tags: [admin]
      security:
        - AdminSession: []
      requestBody:
        required: true
        content:
          application/octet-stream:
            schema:
              type: string
              description: Raw YAML content
      responses:
        '200':
          description: Configuration saved
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
                  message:
                    type: string
                    example: Configuration saved successfully. Restart the server for changes to take effect.
                  path:
                    type: string
        '400':
          description: Invalid YAML syntax
        '401':
          $ref: '#/components/responses/Unauthorized'
  /admin/sessions:
    get:
      operationId: getSessions
      summary: List active MCP sessions
      tags: [admin]
      security:
        - AdminSession: []
      responses:
        '200':
          description: Active MCP sessions
          content:
            application/json:
              schema:
                type: object
                required: [sessions]
                properties:
                  sessions:
                    type: array
                    items:
                      $ref: '#/components/schemas/SessionSnapshot'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /admin/events:
    get:
      operationId: getEventStream
      summary: Live Server-Sent Events stream (audit + agent events)
      description: >-
        Opens an SSE connection streaming real-time audit and agent events.
        Sends "connected" on open, "audit"/"agent" for data, and "ping" keepalives every 30s.
      tags: [admin]
      security:
        - AdminSession: []
      responses:
        '200':
          description: SSE event stream
          content:
            text/event-stream:
              schema:
                type: string
                description: >-
                  Stream of SSE events. Event types: connected, audit, agent, ping.
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ──────────────────────────────────────────────
  # OAuth (Google consent flow helpers)
  # ──────────────────────────────────────────────
  /oauth/start:
    post:
      operationId: oauthStart
      summary: Generate a Google OAuth consent URL
      tags: [oauth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OAuthStartRequest'
      responses:
        '200':
          description: Authorization URL to redirect user to
          content:
            application/json:
              schema:
                type: object
                properties:
                  auth_url:
                    type: string
                    format: uri
                  redirect_uri:
                    type: string
                    format: uri
        '400':
          description: Missing client_id
...
components:
...
  securitySchemes:
    AdminSession:
      type: apiKey
      in: cookie
      name: skyline_admin
      description: >-
        Admin session cookie set by POST /admin/auth. Secure, HttpOnly,
        SameSite=Strict, 7-day expiry.

    ProfileToken:
      type: http
      scheme: bearer
      description: >-
        Per-profile bearer token. Each profile has its own token used for
        authentication when accessing profile data, tools, and MCP endpoints.

  # ──────────────────────────────────────────────
  # Parameters
  # ──────────────────────────────────────────────
...
  schemas:
...
    DetectRequest:
      type: object
      required: [base_url]
      properties:
        base_url:
          type: string
          format: uri
          description: Base URL to probe for API specs
        bearer_token:
          type: string
          description: Optional bearer token forwarded during probing
    DetectResponse:
      type: object
      required: [base_url, online, detected]
      properties:
        base_url:
          type: string
        online:
          type: boolean
          description: True if at least one probe succeeded
        detected:
          type: array
          items:
            $ref: '#/components/schemas/DetectProbe'
...